+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2
 
-- When `DefaultMap` missing key accessed, it also creates an entry in the map. Closes [`#1`](https://github.com/dartsidedev/pop/issues/1).
+- When `DefaultMap` missing key accessed, it also creates an entry in the map [`#1`](https://github.com/dartsidedev/pop/issues/1)
+- Add `entries` extension with index and value to `List` [`#2`](https://github.com/dartsidedev/pop/issues/2)
 
 ## 0.1.1